Restructure code into src dir
